requirements.txt
appdirs==1.4.4
asgiref==3.3.1
auditlog3==1.0.1
boto3==1.16.48
botocore==1.19.48
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-jet==1.0.8
django-storages==1.11.1
django-summernote==0.8.11.6
django-tinymce==3.2.0
docutils==0.16
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
jet-python==0.2
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
virtualenv==20.2.2
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
